Feature/authorize movies and seances (#14)
* add pundit

* add movie policy and update view&controller

* add seance policy and update view&controller

* ref:apply rule to all methods at once

* ref:apply rule to all methods at once

* test:update specs

* chore:disable I18nLocaleTexts cop

+++ app/controllers/movies_controller.rb
@@ -1,18 +1,24 @@
 class MoviesController < ApplicationController
+  before_action :authenticate_user!, except: [:index]
+
   def index
     render :index, locals: { movies: Movie.all }
   end
 
   def new
+    authorize Movie
     render :new, locals: { movie: Movie.new }
   end
 
   def edit
+    authorize movie
     render :edit, locals: { movie: }
   end
 
   def create
     record = Movie.new(movie_params)
+
+    authorize record
 
     if record.save
       redirect_to movies_path, notice: t('.notice')
@@ -22,6 +28,8 @@ class MoviesController < ApplicationController
   end
 
   def update
+    authorize movie
+
     if movie.update(movie_params)
       redirect_to movies_path, notice: t('.notice')
     else
@@ -30,6 +38,8 @@ class MoviesController < ApplicationController
   end
 
   def destroy
+    authorize movie
+
     movie.destroy
     redirect_to movies_path, notice: t('.notice')
   end
